chore: Move Tabs component to UI directory
