--- packages/webapp-libs/webapp-core/src/components/tabs/index.ts
@@ -1,4 +1,0 @@
-export * from './tabs.component';
-export * from './tabsContent';
-export * from './tabsList';
-export * from './tabsTrigger';
