--- packages/webapp-libs/webapp-core/src/components/tabs/tabs.component.tsx
@@ -1,5 +1,0 @@
-import * as TabsPrimitive from '@radix-ui/react-tabs';
-
-const Tabs = TabsPrimitive.Root;
-
-export { Tabs };
