--- packages/webapp-libs/webapp-core/src/components/tabs/tabsContent/index.ts
@@ -1,1 +1,0 @@
-export * from './tabsContent.component';
