--- packages/webapp-libs/webapp-core/src/components/tabs/tabsContent/tabsContent.component.tsx
@@ -1,20 +1,0 @@
-import * as TabsPrimitive from '@radix-ui/react-tabs';
-import * as React from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const TabsContent = React.forwardRef<
-  React.ElementRef<typeof TabsPrimitive.Content>,
-  React.ComponentPropsWithoutRef<typeof TabsPrimitive.Content>
->(({ className, ...props }, ref) => (
-  <TabsPrimitive.Content
-    ref={ref}
-    className={cn(
-      'ring-offset-background focus-visible:ring-ring mt-2 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2',
-      className
-    )}
-    {...props}
-  />
-));
-TabsContent.displayName = TabsPrimitive.Content.displayName;
-export { TabsContent };
