--- packages/webapp-libs/webapp-core/src/components/tabs/tabsList/index.ts
@@ -1,1 +1,0 @@
-export * from './tabsList.component';
