--- packages/webapp-libs/webapp-core/src/components/tabs/tabsTrigger/index.ts
@@ -1,1 +1,0 @@
-export * from './tabsTrigger.component';
