+++ packages/webapp-libs/webapp-finances/src/routes/subscriptions/paymentMethod.content.tsx
@@ -2,8 +2,8 @@ import { useQuery } from '@apollo/client';
 import { getFragmentData } from '@sb/webapp-api-client/graphql';
 import { Link } from '@sb/webapp-core/components/buttons';
 import { PageHeadline } from '@sb/webapp-core/components/pageHeadline';
-import { TabsContent } from '@sb/webapp-core/components/tabs';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@sb/webapp-core/components/ui/card';
+import { TabsContent } from '@sb/webapp-core/components/ui/tabs';
 import { mapConnection } from '@sb/webapp-core/utils/graphql';
 import { useGenerateTenantPath } from '@sb/webapp-tenants/hooks';
 import { useCurrentTenant } from '@sb/webapp-tenants/providers';
